Revert "Dynamic price lookups"
