+++ StockProject/src/main/java/com/example/StockTotalObject.java
