+++ StockProject/src/main/webapp/WEB-INF/jsps/stockinfo.jsp
