+++ StockProject/src/main/webapp/WEB-INF/jsps/stocks.jsp
